% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdb.R
\docType{data}
\name{am_countries}
\alias{am_countries}
\title{Dataset of Countries in the Database}
\format{
A data frame with 55 rows and 9 variables:
\describe{
  \item{Country}{Short Country Name}
  \item{Country_ISO}{ISO Standardized Country Name}
  \item{ISO2}{ISO 2-Character Country Code}
  \item{ISO3}{ISO 3-Character Country Code}
  \item{ISO3N}{ISO Numeric Country Code}
  \item{IMF}{IMF Numeric Country Code}
  \item{Region}{2-Region Classification (UN except for Sudan)}
  \item{Region_Detailed}{5-Region Classification (Former World Bank)}
  \item{Currency}{Main Official Currency}
}
}
\source{
\href{https://CRAN.R-project.org/package=countrycode}{countrycode} R package (with some modification of regional aggregates).
}
\usage{
am_countries
}
\description{
A dataset containing standardized codes of African countries
for which data is provided according to various classifications
and regional aggregations. Note that the API (\code{\link{am_data}})
only supports "ISO3" character codes.
}
\examples{
head(am_countries)
}
\seealso{
\code{\link{am_entities}}, \code{\link{am_sources}}, \code{\link{am_series}}, \code{\link{africamonitor}}
}
\keyword{datasets}
