% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdb.R
\docType{data}
\name{am_entities}
\alias{am_entities}
\title{Dataset of African Economic and Regional Entities}
\format{
A data frame with 54 rows (one for each country, excluding Western Sahara) and 27 variables, of which 6 are country identifiers and the remaining 21
are logical variables indicating country membership to various economic and regional entities.
}
\source{
Own compilation.
}
\usage{
am_entities
}
\description{
A dataset mapping African countries to various economic and regional entities.
}
\examples{
head(am_entities)
}
\seealso{
\code{\link{am_countries}}, \code{\link{am_sources}}, \code{\link{am_series}}, \code{\link{africamonitor}}
}
\keyword{datasets}
