% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdb.R
\docType{data}
\name{.AMID}
\alias{.AMID}
\alias{.AMT}
\title{Global Identifier Macros}
\usage{
.AMID
.AMT
}
\description{
The macro \code{.AMID} contains the string \code{c("ISO3", "Series")} denoting
the names of ID variables that identify the cross-sectional dimension in the database.

The macro \code{.AMT} contains the string
\code{c("Date", "Year", "Quarter", "FY", "QFY", "Month", "Day")} denoting temporal identifiers generated by \code{\link{am_expand_date}}.
The "Date" variable is sufficient to uniquely identify a point in time in the database.

Each value in the database is uniquely identified by ISO3, Series and Date.
}
\examples{
.AMID
.AMT


}
\seealso{
\code{\link{africamonitor}}
}
\keyword{datasets}
